removed all irrelevant code for now
