+++ cucu_train/cucuTrain.py
@@ -12,9 +12,9 @@ import re
 import time
 import numpy as np
 import cv2
-
-
 from PIL import Image
+from cucu_utils import *
+
 
 # Root directory of the project
 ROOT_DIR = os.path.abspath("../")
@@ -90,165 +90,6 @@ class ShapesConfig(Config):
     
 config = ShapesConfig()
 config.display()
-
-
-
-# In[17]:
-
-
-
-
-def get_ax(rows=1, cols=1, size=8):
-    """Return a Matplotlib Axes array to be used in
-    all visualizations in the notebook. Provide a
-    central point to control graph sizes.
-    
-    Change the default size attribute to control the size
-    of rendered images
-    """
-    _, ax = plt.subplots(rows, cols, figsize=(size*cols, size*rows))
-    return ax
-
-
-
-# In[18]:
-
-
-
-
-def rotate_bound(image, angle):
-    '''
-    input:
-        image - image with attribute shape contating height and width
-        angle - rotation angle in degrees
-    
-    returns rotated image.
-    '''
-    # gggrab the dimensions of the image and then determine the
-    # center
-    # first 2 avlues of shape are height,width
-    (h, w) = image.shape[:2]
-    (cX, cY) = (w // 2, h // 2)
- 
-    # grab the rotation matrix (applying the negative of the
-    # angle to rotate clockwise), then grab the sine and cosine
-    # (i.e., the rotation components of the matrix)
-    M = cv2.getRotationMatrix2D((cX, cY), -angle, 1.0)
-    cos = np.abs(M[0, 0])
-    sin = np.abs(M[0, 1])
- 
-    # compute the new bounding dimensions of the image
-    nW = int((h * sin) + (w * cos))
-    nH = int((h * cos) + (w * sin))
- 
-    # adjust the rotation matrix to take into account translation
-    M[0, 2] += (nW / 2) - cX
-    M[1, 2] += (nH / 2) - cY
- 
-    # perform the actual rotation and return the image
-    return cv2.warpAffine(image, M, (nW, nH))
-
-
-def add_image(img1, img2, x_center, y_center, x_scale, y_scale, angle):
-    """
-    name I gave: pasteImageOnOther ruined the overriding of base class function
-    pasting re-scaled image on other image (collage effect)
-    """
-    img2 = img2.resize((int(x_scale*img2.size[0]),int(y_scale*img2.size[1])), Image.ANTIALIAS)
-
-    img2 = img2.rotate(angle, resample=Image.BICUBIC, expand=True)
-    
-
-    rows,cols,channels = np.asarray(img2).shape
-    x_from = x_center - math.floor(cols/2.)
-    y_from = y_center - math.floor(rows/2.)
-
-    img1.paste(img2, (x_from, y_from), img2)
-
-    return img1
-
-def add_imageWithoutTransparency(img1, img2, x_center, y_center, x_scale, y_scale, angle):
-    """
-    pasting re-scaled image on other image (collage effect) without transparency
-    """
-    
-    img2 = cv2.resize(img2,None,fx=x_scale, fy=y_scale, interpolation = cv2.INTER_CUBIC)
-
-    #Rotate
-    img2 = rotate_bound(img2, 360-angle)
-    
-    # I want to put logo on top-left corner, So I create a ROI
-    rows,cols,channels = img2.shape
-    x_from = x_center - math.floor(cols/2.)
-    x_to = x_center + math.ceil(cols/2.)
-    y_from = y_center - math.floor(rows/2.)
-    y_to = y_center + math.ceil(rows/2.)
-    
-    y_max, x_max, _ = img1.shape
-    
-    #cases image out of bg image
-    if x_from < 0:
-        img2 = img2[:,-x_from:]
-        x_from = 0
-    if x_to >= x_max:
-        img2 = img2[:,:-(x_to-x_max+1)]
-        x_to = x_max-1
-    if y_from < 0:
-        img2 = img2[-y_from:,:]
-        y_from = 0
-    if y_to >= y_max:
-        img2 = img2[:-(y_to-y_max+1),:]
-        y_to = y_max-1
-    
-    roi = img1[y_from:y_to, x_from:x_to]
-
-    # Now create a mask of logo and create its inverse mask also
-    alpha_image = mask_from_RGBA(img2)
-    
-    # dilate the mask
-    # kernel = np.ones((5,5),np.uint8)
-    # alpha_image = cv2.dilate(alpha_image, kernel, iterations=1)
-    
-    # since alpha values are in range [0,1] and we know that transparent bg has alpha < 0.4 approx.    
-    _ , mask = cv2.threshold(alpha_image, 0.4 , 255, cv2.THRESH_BINARY)
-    
-    mask_inv = cv2.bitwise_not(mask)
-    # Now black-out the area of logo in ROI
-    img1_bg = cv2.bitwise_and(roi,roi,mask = mask_inv)
-    # Take only region of logo from logo image.
-    img2_fg = cv2.bitwise_and(img2,img2,mask = mask)
-    # Put logo in ROI and modify the main image
-    #print(img1.shape)
-    #print(img2.shape)
-    dst = cv2.add(img1_bg,img2_fg[:,:,0:3])
-    img1[y_from:y_to, x_from:x_to] = dst
-    return img1
-
-def mask_from_RGBA(image):
-    return image[:,:,-1]
-    
-def image_to_mask(image):
-    mask = np.zeros_like(image)
-    while not np.any(mask):
-        color = (0, 0, 0)
-        distance = np.sum(image, axis=-1)
-        mask = distance > 0
-    return mask
-
-def mask_to_image(mask):
-    x,y,z = mask.shape
-    image = np.zeros((x,y))
-    for i in range(0,z):
-        image += mask[:,:,i]*(i+1)
-    print(z)
-    return image
-
-
-
-# In[19]:
-
-
-
 
 
 
@@ -577,67 +418,6 @@ dataset_val.load_shapes(10, config.IMAGE_SHAPE[0], config.IMAGE_SHAPE[1])
 dataset_val.prepare()
 
 
-
-# In[21]:
-
-
-
-# asher note: this section is irrelevant for training on server
-
-#show n random image&mask train examples
-# n = 5
-# image_ids = np.random.choice(dataset_train.image_ids, n)
-# for image_id in image_ids:
-#     image = dataset_train.load_image(image_id)
-#     mask, class_ids = dataset_train.load_mask(image_id)
-#     visualize.display_top_masks(image, mask, class_ids, dataset_train.class_names, 1)
-
-
-
-# In[24]:
-
-# asher note: this section is irrelevant for training on server
-# w = 16
-# h = 16
-
-# n = 1
-# image_ids = np.random.choice(dataset_train.image_ids, n)
-# for image_id in image_ids:
-#     image = dataset_train.load_image(image_id)
-#     mask, class_ids = dataset_train.load_mask(image_id)
-    
-#     fig = plt.figure(frameon=False, dpi=64)
-#     fig.set_size_inches(w,h)
-
-#     ax = plt.Axes(fig, [0., 0., 1., 1.])
-#     ax.set_axis_off()
-#     fig.add_axes(ax)
-
-#     plt.imshow(image)
-#     fig.savefig(ROOT_DIR + '/cucu_train/peekOnDatasetSamples/' + str(image_id) + '.png') 
-    
-    
-#     fig = plt.figure(frameon=False, dpi=64)
-#     fig.set_size_inches(w,h)
-
-#     ax = plt.Axes(fig, [0., 0., 1., 1.])
-#     ax.set_axis_off()
-#     fig.add_axes(ax)
-    
-#     plt.imshow(mask_to_image(mask))
-#     fig.savefig(ROOT_DIR + '/cucu_train/samples/' + str(image_id) + '.png')
-
-#     plt.show()
-    
-
-
-
-
-# In[11]:
-
-
-
-
 # Create model in training mode
 print(MODEL_DIR)
 model = modellib.MaskRCNN(mode="training", config=config, model_dir=MODEL_DIR)
@@ -676,7 +456,7 @@ elif init_with == "last":
 # Passing layers="heads" freezes all layers except the head
 # layers. You can also pass a regular expression to select
 # which layers to train by name pattern.
-#model.train(dataset_train, dataset_val, learning_rate=config.LEARNING_RATE, epochs=1, layers='heads')
+model.train(dataset_train, dataset_val, learning_rate=config.LEARNING_RATE, epochs=1, layers='heads')
 
 
 
@@ -684,14 +464,12 @@ elif init_with == "last":
 
 
 
-
-model.train(dataset_train, dataset_val, learning_rate=config.LEARNING_RATE / 5, epochs=1, layers="all")
+# asher todo: uncomment later when heads training is working
+# model.train(dataset_train, dataset_val, learning_rate=config.LEARNING_RATE / 5, epochs=1, layers="all")
 
 
 
 # In[15]:
-
-
 
 
 # Save weights
@@ -699,168 +477,4 @@ model.train(dataset_train, dataset_val, learning_rate=config.LEARNING_RATE / 5, 
 # Uncomment to save manually
 model_path = os.path.join(MODEL_DIR, "cucuWheights.h5")
 model.keras_model.save_weights(model_path)
-
-
-
-# In[16]:
-
-
-
-
-class InferenceConfig(ShapesConfig):
-    GPU_COUNT = 1
-    IMAGES_PER_GPU = 1
-
-inference_config = InferenceConfig()
-
-# Recreate the model in inference mode
-model = modellib.MaskRCNN(mode="inference", config=inference_config, model_dir=MODEL_DIR)
-
-# Get path to saved weights
-# Either set a specific path or find last trained weights
-# model_path = os.path.join(ROOT_DIR, ".h5 file name here")
-#model_path = model.find_last()
-
-# asher note: adding here: by_name=True, solved
-# a mismatch between weights and layers - peep later on it
-model.load_weights(filepath=MODEL_DIR + '/cucuWheights.h5',by_name=True)
-
-# asher todo: resolve it later, why Dima made another loading
-
-# # Load trained weights
-# print("Loading weights from ", model_path)
-# model.load_weights(model_path, by_name=True)
-
-
-
-# In[17]:
-
-
-
-
-# Test on a random image
-image_id = random.choice(dataset_val.image_ids)
-print(image_id)
-#image_id = 1
-original_image, image_meta, gt_class_id, gt_bbox, gt_mask = modellib.load_image_gt(dataset_val, inference_config, image_id, use_mini_mask=False)
-
-log("original_image", original_image)
-log("image_meta", image_meta)
-log("gt_class_id", gt_class_id)
-log("gt_bbox", gt_bbox)
-log("gt_mask", gt_mask)
-
-# asher note: irrelevant in server mode
-# visualize.display_instances(original_image, gt_bbox, gt_mask, gt_class_id, dataset_train.class_names, figsize=(8, 8))
-
-
-
-# In[18]:
-
-
-
-#asher todo: put here random valid cucumbers img path to test net
-
-t = cv2.cvtColor(cv2.imread(ROOT_DIR +'/cucu_train/simple_test/test1.jpeg'), cv2.COLOR_BGR2RGB)
-#original_image.shape
-results = model.detect([t], verbose=1)
-
-r = results[0]
-visualize.display_instances(t, r['rois'], r['masks'], r['class_ids'], dataset_train.class_names, r['scores'], ax=get_ax())
-# visualize.save_instances(t, r['rois'], r['masks'], r['class_ids'], dataset_train.class_names, r['scores'], ax=get_ax(), save_to='/media/global/gip-main/data/Dmitry/temp/result_0150_bananas.png')
-t= dataset_train.class_names
-print(t)
-
-
-
-# In[19]:
-
-
-
-
-#from os import walk
-#from os import listdir
-
-# asher todo: figure out later what it does
-
-# model = modellib.MaskRCNN(mode="inference", config=inference_config, model_dir=MODEL_DIR)
-# # Recreate the model in inference mode
-# mypath = '/home/master/Work/logs/shapes20180826T1036/'
-# mypath_out = '/home/master/Work/logs/temp_out/'
-
-# # Get path to saved weights
-# # Either set a specific path or find last trained weights
-# # model_path = os.path.join(ROOT_DIR, ".h5 file name here")
-# #model_path = model.find_last()
-
-
-# t = cv2.cvtColor(cv2.imread('/home/master/Work/Tensorflow/Project07 - MaskRCNN/data/results/N09/avocado/2018-06-21_AV_leaves_01_1024_02.jpg'), cv2.COLOR_BGR2RGB)
-# #original_image.shape
-
-
-
-# #f = []
-# #for (dirpath, dirnames, filenames) in walk(mypath):
-# #    print(os.path.join(mypath,filenames))
-#     #f.extend(filenames)
-#     #break
-    
-# for filename in sorted(os.listdir(mypath)):
-    
-#     full_name = os.path.join(mypath,filename)
-#     model_path = model.find_by_name(full_name)
-#     # Load trained weights
-#     print("Loading weights from ", model_path)
-#     model.load_weights(model_path, by_name=True)
-    
-#     results = model.detect([t], verbose=1)
-
-#     r = results[0]
-#     base = os.path.splitext(filename)[0]
-#     image_name = base + ".png"
-    
-#     visualize.save_instances(t, r['rois'], r['masks'], r['class_ids'], dataset_train.class_names, r['scores'], ax=get_ax(), save_to=os.path.join(mypath_out, image_name))
-
-
-
-# In[20]:
-
-
-results = model.detect([original_image], verbose=1)
-
-r = results[0]
-visualize.display_instances(original_image, r['rois'], r['masks'], r['class_ids'], dataset_val.class_names, r['scores'], ax=get_ax())
-
-
-
-# In[21]:
-
-
-
-#asher todo: use this information later when relevant for fine tuning
-
-# # Compute VOC-Style mAP @ IoU=0.5
-# # Running on 10 images. Increase for better accuracy.
-# image_ids = np.random.choice(dataset_val.image_ids, 100)
-# APs = []
-# for image_id in image_ids:
-#     # Load image and ground truth data
-#     image, image_meta, gt_class_id, gt_bbox, gt_mask =        modellib.load_image_gt(dataset_val, inference_config,
-#                                image_id, use_mini_mask=False)
-#     molded_images = np.expand_dims(modellib.mold_image(image, inference_config), 0)
-#     # Run object detection
-#     results = model.detect([image], verbose=0)
-#     r = results[0]
-#     # Compute AP
-#     AP, precisions, recalls, overlaps =        utils.compute_ap(gt_bbox, gt_class_id, gt_mask,
-#                          r["rois"], r["class_ids"], r["scores"], r['masks'])
-#     APs.append(AP)
-    
-# print("mAP: ", np.mean(APs))
-
-
-# In[ ]:
-
-
-
